LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/mathml/presentation-markup/mrow/inferred-mrow-stretchy-expected.txt


FAIL operator stretching inside Mrow assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Sqrt assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Style assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Error assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Phantom assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Math assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Menclose assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Mpadded assert_greater_than_equal: expected a number greater than or equal to 100 but got 10
FAIL operator stretching inside Unknown assert_greater_than_equal: expected a number greater than or equal to 100 but got 0
↨

↨

↨

↨


↨

↨

↨
